fix: move isTestToolEnabledProject into vscode-extension
