+++ packages/vscode-extension/test/localdebug/commonUtils.test.ts
@@ -55,4 +55,24 @@ describe("[debug > commonUtils]", () => {
       chai.expect(result).equals("testId");
     });
   });
+
+  describe("isTestToolEnabledProject", () => {
+    const sandbox = sinon.createSandbox();
+
+    afterEach(async () => {
+      sandbox.restore();
+    });
+
+    it("test tool yaml exist", async () => {
+      sandbox.stub(fs, "pathExistsSync").returns(true);
+      const res = commonUtils.isTestToolEnabledProject("testPath");
+      chai.assert.isTrue(res);
+    });
+
+    it("test tool yaml not exist", async () => {
+      sandbox.stub(fs, "pathExistsSync").returns(false);
+      const res = commonUtils.isTestToolEnabledProject("testPath");
+      chai.assert.isFalse(res);
+    });
+  });
 });
